refactor: do not create copy of array if it is not required
Old browser, including phantomjs 1.X need a "real" as
the context Array.prototype function (it means that
array-like are not permitted).

For exemple:
  var obj = {};
  obj[0] = 'foo';

  // It is not permitted, it need a real array
  Array.prototype.toString.call(obj);

For old browser, an entire copy of array-like object
is created to use a real array object, for other browsers,
keep it fast and do not use a copy.