fix: fix adding elements at the same index
Suppose you try to add elements using `unshift` method:

    collection.unshift(o1);
    // Do some works
    collection.unshift(o2);

It will produce two changes:

- First:
    {
      type: 'splice',
      index: 0,
      addedCount: 1,
      removed: [],
      object: collection
    }

- Second is exactly the same:
    {
      type: 'splice',
      index: 0,
      addedCount: 1,
      removed: [],
      object: collection
    }

It means that is is really hard for an observer to handle
both changes (and this is a simple case). A solution could
be to deliver pending changes synchronously, but it seems
a very dirty workaround.

So a new field has been added to change object: array of
added elements when the change occurs. This is not the
official API for `Object.observe`, so I am still thinking
about a better solution (with performance and simplicity
in mind).

It may be removed later, so client code **should not**
rely on this.

This addition has been proposed and can be tracked
here on ecmascript-object-observe repository, issue 24.